refactoring, added test
