+++ core/src/main/java/com/bc/fiduceo/reader/slstr/utility/Oblique500mTransform.java
@@ -1,4 +1,4 @@
-package com.bc.fiduceo.reader.slstr;
+package com.bc.fiduceo.reader.slstr.utility;
 
 import com.bc.fiduceo.core.Dimension;
 import com.bc.fiduceo.core.Interval;
