+++ core/src/test/java/com/bc/fiduceo/reader/slstr/utility/Oblique500mTransformTest.java
@@ -1,7 +1,8 @@
-package com.bc.fiduceo.reader.slstr;
+package com.bc.fiduceo.reader.slstr.utility;
 
 import com.bc.fiduceo.core.Dimension;
 import com.bc.fiduceo.core.Interval;
+import com.bc.fiduceo.reader.slstr.utility.Oblique500mTransform;
 import com.bc.fiduceo.util.NetCDFUtils;
 import org.junit.Test;
 import ucar.ma2.Array;
